+++ ChangeLog.md
@@ -27,6 +27,8 @@
 * Libs/Rust **(Breaking)**: Remove `settings` parameter from `EnvironmentIn::new`
 * Libs/Go **(Breaking)**: Rename `Statistics.AggregateAppStats` to `AggregateAppStatsWithOptions`;
   the old name is used for a version of the method without the `PostOptions`, like elsewhere
+* Libs/Go **(Breaking)**: Change the type of `Message.CreateWithOptions`' parameter `options` from
+  `PostOptions` to `MessageCreateOptions` - this type has an extra boolean `WithContent` field
 * Libs/Go: Add `Authentication.ExpireAll` (and `ExpireAllWithOptions`)
 
 ## Version 1.56.0